Remove unneed intermediate component.  Improve nullness checks
